// src/lib/database.types.ts
export type Database = {
  // Allows to automatically instanciate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
      farms: {
        Row: {
          breeder_name: string | null
          contact_facebook: string | null
          contact_instagram: string | null
          contact_line: string | null
          contact_wechat: string | null
          contact_whatsapp: string | null
          created_at: string | null
          id: number
          information: string | null
          logo_url: string | null
          name: string
          updated_at: string | null
          user_id: string
        }
        Insert: {
          breeder_name?: string | null
          contact_facebook?: string | null
          contact_instagram?: string | null
          contact_line?: string | null
          contact_wechat?: string | null
          contact_whatsapp?: string | null
          created_at?: string | null
          id?: number
          information?: string | null
          logo_url?: string | null
          name: string
          updated_at?: string | null
          user_id: string
        }
        Update: {
          breeder_name?: string | null
          contact_facebook?: string | null
          contact_instagram?: string | null
          contact_line?: string | null
          contact_wechat?: string | null
          contact_whatsapp?: string | null
          created_at?: string | null
          id?: number
          information?: string | null
          logo_url?: string | null
          name?: string
          updated_at?: string | null
          user_id?: string
        }
        Relationships: []
      }
// ...
      products: {
        Row: {
          created_at: string | null
          deleted_at: string | null
          description: string | null
          farm_id: number
          id: number
          image_urls: string[] | null
          name: string
          price: number | null
          product_id: string | null
          sex: string | null
          status: string
          updated_at: string | null
          user_id: string
          year: string | null
        }
        Insert: {
          created_at?: string | null
          deleted_at?: string | null
          description?: string | null
          farm_id: number
          id?: never
          image_urls?: string[] | null
          name: string
          price?: number | null
          product_id?: string | null
          sex?: string | null
          status?: string
          updated_at?: string | null
          user_id: string
          year?: string | null
        }
        Update: {
          created_at?: string | null
          deleted_at?: string | null
          description?: string | null
          farm_id?: number
          id?: never
          image_urls?: string[] | null
          name?: string
          price?: number | null
          product_id?: string | null
          sex?: string | null
          status?: string
          updated_at?: string | null
          user_id?: string
          year?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "products_farm_id_fkey"
            columns: ["farm_id"]
            isOneToOne: false
            referencedRelation: "farms"
            referencedColumns: ["id"]
          },
        ]
      }
    }
// ...
  }
}